% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getApin}
\alias{getApin}
\title{Get analog pin}
\usage{
getApin(pin)
}
\arguments{
\item{pin}{the number of the pin to get (integer)}
}
\value{
the value of the pin.
}
\description{
Get the value of an analog pin
}
\examples{
\dontrun{
rduinoConnect()
# set position of servo to position of potentiometer
off<-getDpin(4)
while (!off) 
{
  angle<-getApin(5)
  angle<- 1.68 * angle + 575
  setServo(9,angle)
  off<-getDpin(4)
}
offServo()

rduinoClose()
}

}
