% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDpin}
\alias{getDpin}
\title{Get digital pin}
\usage{
getDpin(pin)
}
\arguments{
\item{pin}{the number of the pin to get (integer)}
}
\value{
the binary value of the pin.
}
\description{
Get the value of a digital pin
}
\examples{
\dontrun{
rduinoConnect()
# LED remains on until button is pressed
setDpin(5,1)
isPressed<-getDpin(4)
while (!isPressed){ isPressed<-getDpin(4) }
setDpin(5,0)
rduinoClose()
}

}
