% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{onServo}
\alias{onServo}
\title{Set servo}
\usage{
onServo(pin, value)
}
\arguments{
\item{pin}{the number of the pin connected to the servo}

\item{value}{value to set for the servo}
}
\description{
Activate a servo and set a value
}
\examples{
\dontrun{
rduinoConnect()
# set position of servo to position of potentiometer
off<-getDpin(4)
while (!off) 
{
  angle<-getApin(5)
  angle<- 1.68 * angle + 575
  setServo(9,angle)
  off<-getDpin(4)
}
offServo()
 
rduinoClose()
}

}
