% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{onSignal}
\alias{onSignal}
\title{On Signal
Generate a square wave, from 1Hz to 31.25kHz, with resolution of 16us}
\usage{
onSignal(freq, dutyCycle)
}
\arguments{
\item{freq}{the frequency of the wave, in Hz}

\item{dutyCycle}{the percentage of the time the wave should spend at HIGH}
}
\description{
On Signal
Generate a square wave, from 1Hz to 31.25kHz, with resolution of 16us
}
