% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rduinoConnect}
\alias{rduinoConnect}
\title{Rduino connect}
\usage{
rduinoConnect(baud = 38400, mode = "n,8,1", upload = TRUE,
  arduino = NULL)
}
\arguments{
\item{baud}{baud rate}

\item{mode}{communication mode}

\item{upload}{if TRUE, upload the ino file to the device}

\item{arduino}{command used to run arduino as a shell command including 
the path

This function does two things - uploads a .ino file to an Arduino, and 
acts as a wrapper for the \code{serialConnection}
function of the serial package. The options for the communication mode
are available via the helpfile for the \code{serialConnection} command.}
}
\description{
Make a serial connection to an Arduino or similar device
}
\examples{
\dontrun{
rduinoConnect()
rduinoClose()
}

}
