% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rduinoSample}
\alias{rduinoSample}
\title{Rduino Sample}
\usage{
rduinoSample(readPin, time)
}
\arguments{
\item{readPin}{the analog pin to read in from}

\item{time}{the length of time (ms) that the Arduino should read data for}
}
\description{
Samples in as quickly as possible from given pin for specified duration
}
