% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setApin}
\alias{setApin}
\title{Set analog pin}
\usage{
setApin(pin, value)
}
\arguments{
\item{pin}{the number of the pin to set (integer)}

\item{value}{the value to which to set the pin (real)}
}
\description{
Set a analog pin to on or off
}
\examples{
\dontrun{
rduinoConnect()
# gradually increase intensity of LED
for (i in seq(1,256,by=5)) 
{
  setApin(11,i)
  Sys.sleep(0.05)
}
rduinoClose()
}

}
