% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setDpin}
\alias{setDpin}
\title{Set digital pin}
\usage{
setDpin(pin, value)
}
\arguments{
\item{pin}{the number of the pin to set (integer)}

\item{value}{the value to which to set the pin (binary)}
}
\description{
Set a digital pin to on or off
}
\examples{
\dontrun{
rduinoConnect()
# flash LED rapidly
for (i in 0:9) 
{
  setDpin(8,1)
  Sys.sleep(0.05)
  setDpin(8,0)
  Sys.sleep(0.05)
}
rduinoClose()
}

}
